requirements.txt
streamlit==1.35.0
pandas==2.1.4
numpy==1.26.3
networkx==3.2.1
matplotlib==3.8.2
plotly==5.18.0
pyvis==0.3.2
gspread==6.0.0
google-auth==2.27.0
python-louvain==0.16
python-dotenv==1.0.0
google-generativeai==0.4.0
Pillow==10.2.0
requests==2.31.0 